+++ app/src/main/jni/handwriting_jni.cc
@@ -155,15 +155,6 @@ Java_com_kingzcheung_xime_handwriting_HandwritingNativeEngine_nativeInitialize(
         return JNI_FALSE;
     }
 
-    status = api->SetIntraOpNumThreads(session_options, 2);
-    if (status) {
-        LOGE("Failed to set intra op num threads: %s", api->GetErrorMessage(status));
-        api->ReleaseStatus(status);
-        api->ReleaseSessionOptions(session_options);
-        env->ReleaseStringUTFChars(model_path, modelPathStr);
-        return JNI_FALSE;
-    }
-
     status = api->DisableCpuMemArena(session_options);
     if (status) {
         LOGE("Failed to disable CPU mem arena: %s", api->GetErrorMessage(status));
