+++ app/src/main/jni/onnx_env.cc
@@ -6,6 +6,11 @@
 static OrtEnv* g_shared_env = nullptr;
 static std::mutex g_env_mutex;
 static bool g_env_created = false;
+static OrtThreadingOptions* g_threading_options = nullptr;
+
+// 共享 intra-op 线程池：所有 ONNX 模型（联想/标点/手写）共用一个 env，
+// 因此共用一个线程池，避免每个 session 各自创建线程导致资源/内存浪费。
+#define ONNX_INTRA_OP_THREADS 2
 
 #define LOG_TAG "OnnxEnv"
 #define LOGD(...) __android_log_print(ANDROID_LOG_DEBUG, LOG_TAG, __VA_ARGS__)
@@ -28,13 +33,25 @@ OrtEnv* OnnxGetSharedEnv() {
     if (!g_shared_env) {
         const OrtApi* api = OnnxGetApi();
         if (api) {
-            OrtStatus* status = api->CreateEnv(ORT_LOGGING_LEVEL_WARNING, "xime_onnx", &g_shared_env);
+            OrtStatus* status = api->CreateThreadingOptions(&g_threading_options);
+            if (!status) {
+                status = api->SetGlobalIntraOpNumThreads(g_threading_options, ONNX_INTRA_OP_THREADS);
+            }
+            if (!status) {
+                status = api->CreateEnvWithGlobalThreadPools(
+                    ORT_LOGGING_LEVEL_WARNING, "xime_onnx", g_threading_options, &g_shared_env);
+            }
             if (status) {
                 LOGE("Failed to create shared env: %s", api->GetErrorMessage(status));
                 api->ReleaseStatus(status);
+                if (g_threading_options) {
+                    api->ReleaseThreadingOptions(g_threading_options);
+                    g_threading_options = nullptr;
+                }
             } else {
                 g_env_created = true;
-                LOGD("Onnx shared env created");
+                LOGD("Onnx shared env created (global thread pool, %d intra-op threads)",
+                     ONNX_INTRA_OP_THREADS);
             }
         }
     }
@@ -43,14 +60,16 @@ OrtEnv* OnnxGetSharedEnv() {
 
 void OnnxReleaseSharedEnv() {
     std::lock_guard<std::mutex> lock(g_env_mutex);
+    const OrtApi* api = OnnxGetApi();
     if (g_shared_env) {
-        const OrtApi* api = OnnxGetApi();
-        if (api) {
-            api->ReleaseEnv(g_shared_env);
-            g_shared_env = nullptr;
-            g_env_created = false;
-            LOGD("Onnx shared env released");
-        }
+        if (api) api->ReleaseEnv(g_shared_env);
+        g_shared_env = nullptr;
+        g_env_created = false;
+        LOGD("Onnx shared env released");
+    }
+    if (g_threading_options) {
+        if (api) api->ReleaseThreadingOptions(g_threading_options);
+        g_threading_options = nullptr;
     }
 }
 
@@ -78,5 +97,29 @@ bool OnnxTryEnableNnapi(OrtSessionOptions* options) {
     }
 
     LOGD("NNAPI execution provider enabled successfully");
+    return true;
+}
+
+bool OnnxTryEnableCpuFallback(OrtSessionOptions* options) {
+    const OrtApi* api = OnnxGetApi();
+    if (!api || !options) return false;
+
+    typedef OrtStatus* (*CpuProviderFn)(OrtSessionOptions*, int);
+    CpuProviderFn cpu_fn = (CpuProviderFn)dlsym(
+        RTLD_DEFAULT, "OrtSessionOptionsAppendExecutionProvider_CPU");
+
+    if (!cpu_fn) {
+        LOGD("CPU EP symbol not found (relying on built-in default fallback)");
+        return false;
+    }
+
+    OrtStatus* status = cpu_fn(options, 1);  // use CPU arena
+    if (status) {
+        LOGW("CPU EP append failed: %s", api->GetErrorMessage(status));
+        api->ReleaseStatus(status);
+        return false;
+    }
+
+    LOGD("CPU execution provider appended as fallback");
     return true;
 }
